podman push may fail, retry it
It seems if 2 tags are getting written on the same registry, and if
those 2 tags are from the same base, then this confuses the registry.

The second attempt at rewriting it works, as now when the `push` is
started, the registry already has the blobs available, and it just
writes the manifest.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>