bootstrap: add a qemu capable base image
It turns out that it is possible to start a VM directly from
an unprivileged container. So provide a new base image that we can
then use to create and launch new qemu images to be able to run
test suites that could not be run in a plain container because they
needed access to the kernel.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>