Get rid of the after_script
There is a change in behavior regarding the skopeo command in the
after script. It now fails, meaning we are not tagging the image
properly.

At first, it seamed like a good idea to have a per-job tag, but
I can't find one single user of it, so I guess this is simply
unneeded anymore, and we should simply not tag with the Job ID.

Related: #3

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>