templates: Set $HOME in build container
During execution of distro commands (e.g. $DEBIAN_EXEC), make sure that
we preserve the value of $HOME for execution in the container.

By default, newer versions of buildah seem to strip the environment
completely. This results in $HOME being empty on entry to the container.

If a build container is later reused for native build steps through the
GitLab Docker executor, $HOME will be preserved. When using things which
depend on the value of $HOME, like Python's pip, we can end up with
different expectations. In particular, pip will install into
$HOME/.local/bin, but since $HOME is $HOME, we can't later find what it
actually installed.

Signed-off-by: Daniel Stone <daniels@collabora.com>