template: container_build: remove useless before_script
The `before_script` in the main container build was not that useful:
- `*fdo_distro_vars` was set but not used (and rewritten in `script`)
- the login was there, which allowed us to not have it in the `script`

However, this means anybody who wants to rewrite a `before_script` needs
to remember to also log in to the registry. Better ensure we do it in the
script so that we do not add burden on users.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>