templates: mount the /cache folder RW
`/cache` is often used to speed up compilation with `ccache`. We should
allow the users of CI templates to mount this directory if it exists
so they can set up their runners as they wish and benefit from `ccache`.

Note that to allow the CI to test that we can mount the `/cache` folder.
So we introduce a new variable `CACHE_DIR`, which defaults to `/cache`,
that allows to tell which folder needs to be mounted as `/cache` in the
`buildah run $DISTRIBUTION_EXEC` stage.
- this allows to overwrite the `/cache` folder if it already exists in
  the current runner
- this also allows to export its content as an artifact in the CI job

Fixes #5

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>