doc: add sphinx docs
This just the most basic scaffolding and build script to hook up to gitlab
pages. Actual docs need to be added.

The build script is a minimal wrapper around the sphinx-build command but this
way we can easily ensure that local rebuilding of the docs produces the same
output that gitlab pages will display.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>