Swap the naming order around in the templates
New convention is: .fdo.<job>@<distribution>[@<arch>]
which puts the thing to do at the most significant place and then the ability
to rotate through distributions at the LSBs.

This patch only affects the template itself, the CI scripts we use in this
repo are left with the same previous naming system - mostly because of
laziness.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>